Move and rename the EdgeProvider class
